Move `(De)serializable` traits to new `monad-traits` crate
The two traits were originally defined in `monad-types`, which depends on
`monad-crypto` crate. The `monad-crypto` crate can use the traits, so lifting
them to a separate crate to avoid circular dependencies.
