+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "monad-state",
     "monad-testutil",
     "monad-tracing-counter",
+    "monad-traits",
     "monad-types",
     "monad-validator",
     "monad-viz",
