+++ monad-p2p/src/behavior/codec.rs
@@ -2,7 +2,7 @@ use std::{io::ErrorKind, marker::PhantomData, sync::Arc};
 
 use async_trait::async_trait;
 use monad_executor::{Message, PeerId};
-use monad_types::{Deserializable, Serializable};
+use monad_traits::{Deserializable, Serializable};
 
 use futures::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
 use libp2p::request_response::Codec;
