+++ monad-p2p/src/lib.rs
@@ -8,7 +8,7 @@ use std::{
 
 use futures::{FutureExt, Stream, StreamExt};
 use monad_executor::{Executor, Message, RouterCommand};
-use monad_types::{Deserializable, Serializable};
+use monad_traits::{Deserializable, Serializable};
 
 use libp2p::{request_response::RequestId, swarm::SwarmBuilder, Transport};
 
@@ -319,7 +319,7 @@ mod tests {
 
     use crate::Service;
     use monad_executor::Message;
-    use monad_types::{Deserializable, Serializable};
+    use monad_traits::{Deserializable, Serializable};
 
     use futures::StreamExt;
