+++ monad-wal/benches/event_bench.rs
@@ -31,7 +31,7 @@ use monad_testutil::{
     block::setup_block,
     signing::{create_keys, get_key},
 };
-use monad_types::Serializable;
+use monad_traits::Serializable;
 use monad_types::{BlockId, Hash, NodeId, Round};
 use monad_wal::wal::{WALogger, WALoggerConfig};
 use monad_wal::PersistenceLogger;
